[Chore] #68 Protocol 폴더를 생성하여 채움함에서 쓰인 protocol 코드를 이동
채움함의 protocol만 옮긴 이유는
코디네이터 관련 코드로 ~Flow가 protocol이었는데 함부러 옮겨도 될지 모르겠어서
또한 해당 protocol이 Coordinator와 같은 파일에 있는 게 보기에 더 편할 것 같기도 해서 이동하지 않음
