+++ Milestone/Milestone/Global/Support/Coordinator.swift
@@ -6,6 +6,7 @@
 //
 
 import Foundation
+
 protocol Coordinator {
     func start()
     func coordinate(to coordinator: Coordinator)
